Refactor update gui_toolkit.py for resource loading
- Moved the `get_sprite` function to `resource_loader.py` for better organization.
- Created a new function `asset_menu_btn` in `resource_loader.py` to attach an image displayed as a menu button onto the application window.
- Created a new function `update_asset` in `resource_loader.py` to update an asset in a tkinter canvas with the provided image.
- Created a new function `set_custom_btn_styles` in `resource_loader.py` to set custom button styles.
- Moved the `set_ttk_theme` function to `resource_loader.py` for better organization.
